loganly/requirements.txt
PyYAML==5.1.2
SQLAlchemy==1.3.7
